addressing review feedbacks
