+++ core/src/main/java/org/apache/gravitino/GravitinoEnv.java
@@ -68,7 +68,6 @@ import org.apache.gravitino.hook.SchemaHookDispatcher;
 import org.apache.gravitino.hook.TableHookDispatcher;
 import org.apache.gravitino.hook.TagHookDispatcher;
 import org.apache.gravitino.hook.TopicHookDispatcher;
-import org.apache.gravitino.hook.ViewHookDispatcher;
 import org.apache.gravitino.job.BuiltInJobTemplateEventListener;
 import org.apache.gravitino.job.JobManager;
 import org.apache.gravitino.job.JobOperationDispatcher;
@@ -635,13 +634,17 @@ public class GravitinoEnv {
         new FunctionEventDispatcher(eventBus, functionNormalizeDispatcher);
     this.functionDispatcher = new FunctionHookDispatcher(functionEventDispatcher);
 
+    // View operation chain: ViewEventDispatcher -> ViewNormalizeDispatcher ->
+    // ViewOperationDispatcher.
+    // TODO(#11007): Add ViewHookDispatcher for view ownership and privilege hooks when view
+    // privilege support is finalized.
     ViewOperationDispatcher viewOperationDispatcher =
         new ViewOperationDispatcher(catalogManager, entityStore, idGenerator);
     ViewNormalizeDispatcher viewNormalizeDispatcher =
         new ViewNormalizeDispatcher(viewOperationDispatcher, catalogManager);
     ViewEventDispatcher viewEventDispatcher =
         new ViewEventDispatcher(eventBus, viewNormalizeDispatcher);
-    this.viewDispatcher = new ViewHookDispatcher(viewEventDispatcher);
+    this.viewDispatcher = viewEventDispatcher;
 
     this.statisticDispatcher =
         new StatisticEventDispatcher(
