+++ core/src/main/java/org/apache/gravitino/audit/AuditLog.java
@@ -33,8 +33,6 @@ import org.apache.gravitino.listener.api.event.AlterTableEvent;
 import org.apache.gravitino.listener.api.event.AlterTableFailureEvent;
 import org.apache.gravitino.listener.api.event.AlterTopicEvent;
 import org.apache.gravitino.listener.api.event.AlterTopicFailureEvent;
-import org.apache.gravitino.listener.api.event.AlterViewEvent;
-import org.apache.gravitino.listener.api.event.AlterViewFailureEvent;
 import org.apache.gravitino.listener.api.event.CreateCatalogEvent;
 import org.apache.gravitino.listener.api.event.CreateCatalogFailureEvent;
 import org.apache.gravitino.listener.api.event.CreateFilesetEvent;
@@ -47,8 +45,6 @@ import org.apache.gravitino.listener.api.event.CreateTableEvent;
 import org.apache.gravitino.listener.api.event.CreateTableFailureEvent;
 import org.apache.gravitino.listener.api.event.CreateTopicEvent;
 import org.apache.gravitino.listener.api.event.CreateTopicFailureEvent;
-import org.apache.gravitino.listener.api.event.CreateViewEvent;
-import org.apache.gravitino.listener.api.event.CreateViewFailureEvent;
 import org.apache.gravitino.listener.api.event.DisableCatalogEvent;
 import org.apache.gravitino.listener.api.event.DisableCatalogFailureEvent;
 import org.apache.gravitino.listener.api.event.DisableMetalakeEvent;
@@ -65,8 +61,6 @@ import org.apache.gravitino.listener.api.event.DropTableEvent;
 import org.apache.gravitino.listener.api.event.DropTableFailureEvent;
 import org.apache.gravitino.listener.api.event.DropTopicEvent;
 import org.apache.gravitino.listener.api.event.DropTopicFailureEvent;
-import org.apache.gravitino.listener.api.event.DropViewEvent;
-import org.apache.gravitino.listener.api.event.DropViewFailureEvent;
 import org.apache.gravitino.listener.api.event.EnableCatalogEvent;
 import org.apache.gravitino.listener.api.event.EnableCatalogFailureEvent;
 import org.apache.gravitino.listener.api.event.EnableMetalakeEvent;
@@ -91,8 +85,6 @@ import org.apache.gravitino.listener.api.event.ListTableEvent;
 import org.apache.gravitino.listener.api.event.ListTableFailureEvent;
 import org.apache.gravitino.listener.api.event.ListTopicEvent;
 import org.apache.gravitino.listener.api.event.ListTopicFailureEvent;
-import org.apache.gravitino.listener.api.event.ListViewEvent;
-import org.apache.gravitino.listener.api.event.ListViewFailureEvent;
 import org.apache.gravitino.listener.api.event.LoadCatalogEvent;
 import org.apache.gravitino.listener.api.event.LoadCatalogFailureEvent;
 import org.apache.gravitino.listener.api.event.LoadFilesetEvent;
@@ -105,8 +97,6 @@ import org.apache.gravitino.listener.api.event.LoadTableEvent;
 import org.apache.gravitino.listener.api.event.LoadTableFailureEvent;
 import org.apache.gravitino.listener.api.event.LoadTopicEvent;
 import org.apache.gravitino.listener.api.event.LoadTopicFailureEvent;
-import org.apache.gravitino.listener.api.event.LoadViewEvent;
-import org.apache.gravitino.listener.api.event.LoadViewFailureEvent;
 import org.apache.gravitino.listener.api.event.OperationStatus;
 import org.apache.gravitino.listener.api.event.OperationType;
 import org.apache.gravitino.listener.api.event.PartitionExistsEvent;
@@ -115,6 +105,16 @@ import org.apache.gravitino.listener.api.event.PurgePartitionEvent;
 import org.apache.gravitino.listener.api.event.PurgePartitionFailureEvent;
 import org.apache.gravitino.listener.api.event.PurgeTableEvent;
 import org.apache.gravitino.listener.api.event.PurgeTableFailureEvent;
+import org.apache.gravitino.listener.api.event.view.AlterViewEvent;
+import org.apache.gravitino.listener.api.event.view.AlterViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.CreateViewEvent;
+import org.apache.gravitino.listener.api.event.view.CreateViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.DropViewEvent;
+import org.apache.gravitino.listener.api.event.view.DropViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.ListViewEvent;
+import org.apache.gravitino.listener.api.event.view.ListViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.LoadViewEvent;
+import org.apache.gravitino.listener.api.event.view.LoadViewFailureEvent;
 
 /** The interface define unified audit log schema. */
 public interface AuditLog {
