+++ core/src/main/java/org/apache/gravitino/listener/ViewEventDispatcher.java
@@ -28,21 +28,21 @@ import org.apache.gravitino.catalog.ViewOperationDispatcher;
 import org.apache.gravitino.exceptions.NoSuchSchemaException;
 import org.apache.gravitino.exceptions.NoSuchViewException;
 import org.apache.gravitino.exceptions.ViewAlreadyExistsException;
-import org.apache.gravitino.listener.api.event.AlterViewEvent;
-import org.apache.gravitino.listener.api.event.AlterViewFailureEvent;
-import org.apache.gravitino.listener.api.event.AlterViewPreEvent;
-import org.apache.gravitino.listener.api.event.CreateViewEvent;
-import org.apache.gravitino.listener.api.event.CreateViewFailureEvent;
-import org.apache.gravitino.listener.api.event.CreateViewPreEvent;
-import org.apache.gravitino.listener.api.event.DropViewEvent;
-import org.apache.gravitino.listener.api.event.DropViewFailureEvent;
-import org.apache.gravitino.listener.api.event.DropViewPreEvent;
-import org.apache.gravitino.listener.api.event.ListViewEvent;
-import org.apache.gravitino.listener.api.event.ListViewFailureEvent;
-import org.apache.gravitino.listener.api.event.ListViewPreEvent;
-import org.apache.gravitino.listener.api.event.LoadViewEvent;
-import org.apache.gravitino.listener.api.event.LoadViewFailureEvent;
-import org.apache.gravitino.listener.api.event.LoadViewPreEvent;
+import org.apache.gravitino.listener.api.event.view.AlterViewEvent;
+import org.apache.gravitino.listener.api.event.view.AlterViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.AlterViewPreEvent;
+import org.apache.gravitino.listener.api.event.view.CreateViewEvent;
+import org.apache.gravitino.listener.api.event.view.CreateViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.CreateViewPreEvent;
+import org.apache.gravitino.listener.api.event.view.DropViewEvent;
+import org.apache.gravitino.listener.api.event.view.DropViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.DropViewPreEvent;
+import org.apache.gravitino.listener.api.event.view.ListViewEvent;
+import org.apache.gravitino.listener.api.event.view.ListViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.ListViewPreEvent;
+import org.apache.gravitino.listener.api.event.view.LoadViewEvent;
+import org.apache.gravitino.listener.api.event.view.LoadViewFailureEvent;
+import org.apache.gravitino.listener.api.event.view.LoadViewPreEvent;
 import org.apache.gravitino.listener.api.info.ViewInfo;
 import org.apache.gravitino.rel.Column;
 import org.apache.gravitino.rel.Representation;
