+++ core/src/main/java/org/apache/gravitino/listener/api/event/view/ListViewPreEvent.java
@@ -17,11 +17,12 @@
  *  under the License.
  */
 
-package org.apache.gravitino.listener.api.event;
+package org.apache.gravitino.listener.api.event.view;
 
 import org.apache.gravitino.NameIdentifier;
 import org.apache.gravitino.Namespace;
 import org.apache.gravitino.annotation.DeveloperApi;
+import org.apache.gravitino.listener.api.event.OperationType;
 
 /** Pre-event before listing views in a namespace. */
 @DeveloperApi
