+++ core/src/main/java/org/apache/gravitino/listener/api/event/view/ViewFailureEvent.java
@@ -17,33 +17,17 @@
  * under the License.
  */
 
-package org.apache.gravitino.listener.api.event;
+package org.apache.gravitino.listener.api.event.view;
 
 import org.apache.gravitino.NameIdentifier;
 import org.apache.gravitino.annotation.DeveloperApi;
-import org.apache.gravitino.listener.api.info.ViewInfo;
+import org.apache.gravitino.listener.api.event.FailureEvent;
 
-/** Successful load-view event. */
+/** Base class for view operation failure events. */
 @DeveloperApi
-public final class LoadViewEvent extends ViewEvent {
-  private final ViewInfo loadedViewInfo;
+public abstract class ViewFailureEvent extends FailureEvent {
 
-  public LoadViewEvent(String user, NameIdentifier identifier, ViewInfo loadedViewInfo) {
-    super(user, identifier);
-    this.loadedViewInfo = loadedViewInfo;
-  }
-
-  /**
-   * Returns the loaded view information.
-   *
-   * @return the loaded view information
-   */
-  public ViewInfo loadedViewInfo() {
-    return loadedViewInfo;
-  }
-
-  @Override
-  public OperationType operationType() {
-    return OperationType.LOAD_VIEW;
+  protected ViewFailureEvent(String user, NameIdentifier identifier, Exception exception) {
+    super(user, identifier, exception);
   }
 }
